Refactor: [M3-6329] - MUI v5 Migration -
