--- packages/manager/src/components/EditableEntityLabel/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './EditableEntityLabel';
